[Fleet] Support global data tags for agentless (#263522)
(cherry picked from commit 5cc7014c4a0318fb7db325265eaeacfc907b478d)
