+++ x-pack/platform/plugins/shared/fleet/common/types/rest_spec/agentless_policy.ts
@@ -35,6 +35,23 @@ export const CreateAgentlessPolicyRequestSchema = {
         },
       })
     ),
+    // Only available for agentless integration policies.
+    // On standard package policies this field is rejected by server-side validation.
+    global_data_tags: schema.maybe(
+      schema.arrayOf(
+        schema.object({
+          name: schema.string({
+            meta: { description: 'The name of the custom field. Cannot contain spaces.' },
+          }),
+          value: schema.oneOf([schema.string(), schema.number()], {
+            meta: { description: 'The value of the custom field.' },
+          }),
+        }),
+        {
+          maxSize: 100,
+        }
+      )
+    ),
     // Cloud connector configuration - all connector settings go here
     cloud_connector: schema.maybe(
       schema.object({
