+++ x-pack/platform/plugins/shared/fleet/public/applications/fleet/components/custom_fields/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { AgentPolicyCustomFields } from './agent_policy_custom_fields';
+export { PackagePolicyCustomFields } from './package_policy_custom_fields';
+export { GlobalDataTagsTable } from './global_data_tags_table';
